fixing 64 bit

+++ bundles/org.eclipse.swt/Eclipse SWT/cocoa/org/eclipse/swt/widgets/Table.java
@@ -1348,7 +1348,7 @@ void mouseDown (int /*long*/ id, int /*long*/ sel, int /*long*/ theEvent) {
 		NSEvent nsEvent = new NSEvent (theEvent);
 		NSPoint location = nsEvent.locationInWindow ();
 		location = headerView.convertPoint_fromView_ (location, null);
-		int index = headerView.columnAtPoint (location);
+		int index = (int)/*64*/headerView.columnAtPoint (location);
 		if (index != -1) {
 			TableColumn column = columns [index]; // TODO wrong column when columns have been rearranged
 			column.postEvent (SWT.Selection);
